dom/html/test/test_bug660959-3.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=660959
-->
<head>
  <title>Test for Bug 660959</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="reflect.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=660959">Mozilla Bug 660959</a>
<p id="display"></p>
<div id="content" style="display: none">
  <a href="http://www.example.com"></a>
  <div id="foo">
    <span id="test"></span>
  </div>
</div>
<pre id="test">
<script>
    is($("foo").querySelector(":link + * span, :visited + * span"), $("test"),
       "Should be able to find link siblings even in a display:none subtree");
</script>
</pre>
</body>
</html>
